Renamespace dataset provider
